% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{books}
\alias{books}
\title{Data frame of Ming and Qing novels}
\usage{
books()
}
\value{
A data frame with two columns: \code{text} and \code{book}
}
\description{
Returns a data frame of Ming and Qing novels with
only 2 columns: \code{text}, which contains the text,
and \code{book}, which contains the titles.
}
\examples{
library(dplyr)

books() \%>\% group_by(book) \%>\%
    summarise(total_lines = n())

}
