% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{hongloumeng}
\alias{hongloumeng}
\title{The text of Hong Lou Meng}
\format{A character vector with 28409 elements}
\source{
\url{https://www.gutenberg.org/ebooks/24264}
}
\usage{
hongloumeng
}
\description{
The UTF-8 plain text was sourced from Project Gutenberg.
(Some elements might be blank.)
}
\keyword{datasets}
