% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{jingshitongyan}
\alias{jingshitongyan}
\title{The text of Jing Shi Tong Yan}
\format{A character vector with 4386 elements}
\source{
\url{https://www.gutenberg.org/ebooks/24141}
}
\usage{
jingshitongyan
}
\description{
The UTF-8 plain text was sourced from Project Gutenberg.
(Some elements might be blank.)
}
\keyword{datasets}
