% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{package}
\name{mqxsr}
\alias{mqxsr}
\alias{mqxsr-package}
\title{Chinese novels around Ming and Qing}
\description{
This package contains texts in Chinese of some novels
around Ming and Qing dynasties, and perhaps of some
Republic of China titles, from about 15th to early 20th century.
}
