% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{sanguozhiyanyi}
\alias{sanguozhiyanyi}
\title{The text of San Guo Zhi Yan Yi}
\format{A character vector with 21287 elements}
\source{
\url{https://www.gutenberg.org/ebooks/23950}
}
\usage{
sanguozhiyanyi
}
\description{
The UTF-8 plain text was sourced from Project Gutenberg.
(Some elements might be blank.)
}
\keyword{datasets}
