% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{shuihuzhuan}
\alias{shuihuzhuan}
\title{The text of Shui Hu Zhuan}
\format{A character vector with 16900 elements}
\source{
\url{https://www.gutenberg.org/ebooks/23863}
}
\usage{
shuihuzhuan
}
\description{
The UTF-8 plain text was sourced from Project Gutenberg.
(Some elements might be blank.)
}
\keyword{datasets}
