% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{xiyouji}
\alias{xiyouji}
\title{The text of 《西遊記》}
\format{A character vector with 26939 elements}
\source{
\url{https://www.gutenberg.org/ebooks/23962}
}
\usage{
xiyouji
}
\description{
The UTF-8 plain text was sourced from Project Gutenberg.
(Some elements might be blank.)
}
\keyword{datasets}
