% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{yushimingyan}
\alias{yushimingyan}
\title{The text of Yu Shi Ming Yan}
\format{A character vector with 13530 elements}
\source{
\url{https://www.gutenberg.org/ebooks/27582}
}
\usage{
yushimingyan
}
\description{
The UTF-8 plain text was sourced from Project Gutenberg.
(Some elements might be blank.)
}
\keyword{datasets}
